Fixed an infinite number of errors.
Well.  This code is really a mess of C/C++.  It is
a fair sample of the finest workaround technique between
programmers.  I fixed most of syntax errors, some
integer type conflicts, and other I had to ignore by
commenting out because I'm not involved with this project.
But I want to help.  Also it was created a good Makefile
to manage the project.  To check for syntax problems,
type `make check`.  To clean generated files, type
`make clean`.  And to compile, just type `make`.  This last
one will not work because I did not see any main method yet.
Consider my comments, look for C++ style, it will be easier
to adopt C++ rather than mix two styles.
Signed-off-by: Paulo Roberto Urio <jimmyskull@gmail.com>